Refactor nursery withdrawal and report photos services
