+++ src/services/ReportService.ts
@@ -1,6 +1,7 @@
 import { paths } from 'src/api/types/generated-schema';
 import { ListReport, Report, ReportFile, ReportPhoto } from 'src/types/Report';
 import HttpService, { Response } from './HttpService';
+import PhotoService from './PhotoService';
 
 /**
  * Report related services
@@ -28,8 +29,6 @@ type ReportPhotosResponsePayload =
   paths[typeof REPORT_PHOTOS_ENDPOINT]['get']['responses'][200]['content']['application/json'];
 type UploadReportFileResponsePayload =
   paths[typeof UPLOAD_REPORT_FILES_ENDPOINT]['post']['responses'][200]['content']['application/json'];
-type UploadReportPhotoResponsePayload =
-  paths[typeof UPLOAD_REPORT_PHOTO_ENDPOINT]['post']['responses'][200]['content']['application/json'];
 type ReportPhotoResponsePayload = paths[typeof REPORT_PHOTO_ENDPOINT]['get']['responses'][200]['content'];
 
 type UpdateReportRequestPayload = paths[typeof REPORT_ENDPOINT]['put']['requestBody']['content']['application/json'];
@@ -104,7 +103,6 @@ const httpReportPhotos = HttpService.root(REPORT_PHOTOS_ENDPOINT);
 const httpSubmitReport = HttpService.root(SUBMIT_REPORT_ENDPOINT);
 const httpUnlockReport = HttpService.root(UNLOCK_REPORT_ENDPOINT);
 const httpUploadReportFile = HttpService.root(UPLOAD_REPORT_FILES_ENDPOINT);
-const httpUploadReportPhoto = HttpService.root(UPLOAD_REPORT_PHOTO_ENDPOINT);
 const httpReportPhoto = HttpService.root(REPORT_PHOTO_ENDPOINT);
 
 /**
@@ -290,53 +288,11 @@ const submitReport = async (id: number): Promise<Response> => {
 };
 
 /**
- * upload report photo
- */
-const uploadReportPhoto = async (reportId: number, file: File): Promise<UploadReportPhotoResponse> => {
-  const entity = new FormData();
-  entity.append('file', file);
-
-  const headers = {
-    'content-type': 'multipart/form-data',
-  };
-
-  const response: UploadReportPhotoResponse = await httpUploadReportPhoto.post({
-    entity,
-    headers,
-    urlReplacements: {
-      '{reportId}': reportId.toString(),
-    },
-  });
-
-  if (response.requestSucceeded) {
-    const data: UploadReportPhotoResponsePayload = response.data;
-    response.fileId = data?.fileId;
-  }
-
-  return response;
-};
-
-/**
  * Upload multiple photos for a report
  */
 const uploadReportPhotos = async (reportId: number, photos: File[]): Promise<((Response & PhotoId) | string)[]> => {
-  const uploadPhotoPromises = photos.map((photo) => uploadReportPhoto(reportId, photo));
-  try {
-    const promiseResponses = await Promise.allSettled(uploadPhotoPromises);
-    return promiseResponses.map((response) => {
-      if (response.status === 'rejected') {
-        // tslint:disable-next-line: no-console
-        console.error(response.reason);
-        return response.reason;
-      } else {
-        return response.value as Response & PhotoId;
-      }
-    });
-  } catch (e) {
-    // swallow error
-  }
-
-  return [];
+  const url = UPLOAD_REPORT_PHOTO_ENDPOINT.replace('{reportId}', reportId.toString());
+  return PhotoService.uploadPhotos(url, photos);
 };
 
 /**
@@ -397,7 +353,6 @@ const ReportService = {
   unlockReport,
   getReportPhotos,
   submitReport,
-  uploadReportPhoto,
   uploadReportPhotos,
   getReportPhoto,
   updateReportPhoto,
